refactor: unify sqlite schema and db path runtime
