--- polylogue/storage/backends/schema_upgrade.py
@@ -1,70 +1,0 @@
-"""Schema upgrade helpers for the fixed archive schema."""
-
-from __future__ import annotations
-
-import sqlite3
-
-from polylogue.errors import DatabaseError
-from polylogue.logging import get_logger
-from polylogue.storage.backends.schema_bootstrap import (
-    SCHEMA_DDL,
-    SCHEMA_VERSION,
-    apply_schema_extension_plan,
-    assert_supported_archive_layout_snapshot,
-    build_current_schema_extension_plan,
-    capture_schema_snapshot,
-    ensure_vec0_table,
-    schema_version_mismatch_message,
-)
-
-logger = get_logger(__name__)
-
-
-def _ensure_raw_source_mtime_index(conn: sqlite3.Connection) -> None:
-    snapshot = capture_schema_snapshot(conn)
-    plan = build_current_schema_extension_plan(snapshot)
-    for statement in plan.statements:
-        if "idx_raw_conv_source_mtime" in statement:
-            conn.execute(statement)
-
-
-def assert_supported_archive_layout(conn: sqlite3.Connection) -> None:
-    """Reject legacy archive layouts that the current runtime cannot write safely."""
-    assert_supported_archive_layout_snapshot(capture_schema_snapshot(conn))
-
-
-def apply_current_schema_extensions(conn: sqlite3.Connection) -> None:
-    snapshot = capture_schema_snapshot(conn)
-    plan = build_current_schema_extension_plan(snapshot)
-    if snapshot.sql_for_index("idx_raw_conv_source_mtime") is not None and any(
-        statement == "DROP INDEX IF EXISTS idx_raw_conv_source_mtime" for statement in plan.statements
-    ):
-        logger.info("Replacing idx_raw_conv_source_mtime with partial covering definition")
-    apply_schema_extension_plan(conn, plan)
-    ensure_vec0_table(conn)
-    conn.commit()
-
-
-def ensure_schema(conn: sqlite3.Connection) -> None:
-    """Ensure the database is at the current schema version."""
-    snapshot = capture_schema_snapshot(conn)
-
-    if snapshot.current_version == 0:
-        conn.execute("PRAGMA foreign_keys = ON")
-        conn.executescript(SCHEMA_DDL)
-        ensure_vec0_table(conn)
-        conn.execute(f"PRAGMA user_version = {SCHEMA_VERSION}")
-        conn.commit()
-        logger.debug("Created fresh schema v%s", SCHEMA_VERSION)
-        return
-
-    assert_supported_archive_layout_snapshot(snapshot)
-
-    if snapshot.current_version == SCHEMA_VERSION:
-        apply_current_schema_extensions(conn)
-        return
-
-    raise DatabaseError(schema_version_mismatch_message(snapshot.current_version))
-
-
-__all__ = ["apply_current_schema_extensions", "assert_supported_archive_layout", "ensure_schema", "ensure_vec0_table"]
