+++ tests/unit/cli/test_check_runtime.py
@@ -378,14 +378,11 @@ class TestRuntimeHealthLegacySchema:
 
     def test_runtime_health_reports_legacy_inline_raw_layout(self, tmp_path, monkeypatch):
         import polylogue.paths
-        import polylogue.storage.backends.connection as connection_module
-
         monkeypatch.setenv("XDG_DATA_HOME", str(tmp_path / "data"))
         monkeypatch.setenv("XDG_CONFIG_HOME", str(tmp_path / "config"))
         importlib.reload(polylogue.paths)
-        importlib.reload(connection_module)
 
-        db_path = connection_module.default_db_path()
+        db_path = polylogue.paths.db_path()
         db_path.parent.mkdir(parents=True, exist_ok=True)
         conn = sqlite3.connect(db_path)
         conn.executescript(
